Extract SnapshotCardExternalLink layout
